% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csr_graph.R
\name{materialize_csr_edges}
\alias{materialize_csr_edges}
\title{Materialize a compact graph for small-data inspection}
\usage{
materialize_csr_edges(graph, max_edges = 1e+06)
}
\arguments{
\item{graph}{A compact SpatialESS CSR graph.}

\item{max_edges}{Maximum edges allowed for materialization.}
}
\description{
Materialize a compact graph for small-data inspection
}
