% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csr_graph.R
\name{validate_spatial_csr_graph}
\alias{validate_spatial_csr_graph}
\title{Validate a compact SpatialESS CSR graph}
\usage{
validate_spatial_csr_graph(graph)
}
\arguments{
\item{graph}{A compact SpatialESS CSR graph.}
}
\description{
Validate a compact SpatialESS CSR graph
}
